refactor: todosReducer 위치변경
TodoApp에서 reducer를 import 해서 Provider의 인자로 넣어주는것이 아닌, todoState에서 import해서 바로 사용하도록 하여

TodoApp -> todoState로 인자를 넘겨주는 의존관계(?)를 없앰
