+++ src/Header/ResultBar.js
@@ -1,7 +1,7 @@
 import React from "react";
 import RadiusDisplayer from "../components/Displayer";
 import styled from "styled-components";
-import { useStateValue } from "../TodoState";
+import { useStateValue } from "../StateHelper/TodoState";
 
 const ResultWrapper = styled.div`
   display: flex;
