+++ src/Todo/TodoItem.js
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import styled from "styled-components";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faTimes } from "@fortawesome/free-solid-svg-icons";
-import { useStateValue } from "../TodoState";
+import { useStateValue } from "../StateHelper/TodoState";
 
 const Title = styled.span`
   height: 2rem;
